// src/db/entities/Unit/types.ts
/** The values of this enum must exactly match dcs in-game unit typeName values */
export enum UnitTypeName {
  // F1 Supplies and Utility
  M939 = 'M 818', // Truck M939 Heavy
  FPS117 = 'FPS-117', // EWR AN/FPS-117 Radar
  // F2 Light
  HmmwvJtac = 'Hummer', // LUV HMMWV Jeep
  HmmwvMg = 'M1043 HMMWV Armament', // Scout HMMWV
  HmmwvTow = 'M1045 HMMWV TOW', // ATGM HMMWV
  Fuchs = 'TPZ', // APC TPz Fuchs
  LAV25 = 'LAV-25', // IFV LAV-25
  StrykerMgs = 'M1128 Stryker MGS', // SPG Stryker MGS
  StrykerAtgm = 'M1134 Stryker ATGM', // ATGM Stryker
  MephistoAtgm = 'VAB_Mephisto', // ATGM VAB Mephisto
  // F3 Heavy
  M1A2 = 'M-1 Abrams', // MBT M1A2 Abrams
  M2A2 = 'M-2 Bradley', // IFV M2A2 Bradley
  Marder = 'Marder', // IFV Marder
  MCV80 = 'MCV-80', // IFV Warrior
  Leopard2 = 'Leopard-2', // MBT Leopard-2A6M
  Leclerc = 'Leclerc', // MBT Leclerc
  Challenger2 = 'Challenger2', // MBT Challenger II
  // F4 Artillery
  M109 = 'M-109', // SPH M109 Paladin 155mm
  Dana = 'SpGH_Dana', // SPH Dana vz77 152mm
  Firtina = 'T155_Firtina', // SPH T155 Firtina 155mm
  MlrsM270 = 'MLRS', // MLRS M270 227mm
  Mlrs9A52 = 'Smerch_HE', // MLRS 9A52 Smerch HE 300mm
  // F5 - SHORAD
  Vulcan = 'Vulcan', // SPAAA Vulcan M163
  Gepard = 'Gepard', // SPAAA Gepard
  Avenger = 'M1097 Avenger', // SAM Avenger (Stinger)
  Chaparral = 'M48 Chaparral', // SAM Chaparral M48
  RolandAds = 'Roland ADS', // SAM Roland ADS
  Linebacker = 'M6 Linebacker', // SAM Linebacker - Bradley M6
  // F6 - MRAD
  HawkSr = 'Hawk sr', // SAM Hawk SR (AN/MPQ50)*
  HawkCwar = 'Hawk cwar', // SAM Hawk CWAR (AN/MPQ55)
  HawkTr = 'Hawk tr', // SAM Hawk TR (AN/MPQ46)
  HawkPcp = 'Hawk pcp', // SAM Hawk Platoon Command Post (PCP)
  HawkLn = 'Hawk ln', // SAM Hawk LN M192
  NasamLnB = 'NASAMS_LN_B', // SAM NASAMS LN AIM-120B
  NasamLnC = 'NASAMS_LN_C', // SAM NASAMS LN AIM-120C
  NasamSr = 'NASAMS_Radar_MPQ64F1', // SAM NASAMS SR MPQ64F1
  NasamC2 = 'NASAMS_Command_Post', // SAM NASAMS C2
  // helicopters
  UH1H = 'UH-1H',
  AH64D = 'AH-64D_BLK_II',
  MI24 = 'Mi-24P',
  MI8 = 'Mi-8MT',
  UH60 = 'UH-60L',
  KA50 = 'Ka-50',
  KA503 = 'Ka-50_3',
  SA342L = 'SA342L',
  SA342M = 'SA342M',
  SA342Minigun = 'SA342Minigun',
  SA342Mistral = 'SA342Mistral',
  // utility
  M818 = 'M 818',
  M978 = 'M978 HEMTT Tanker',
  // infantry
  InfantryAk = 'Infantry AK',
  InfantryAkRusVer1 = 'Soldier AK',
  InfantryAkRusVer2 = 'Infantry AK ver2',
  InfantryAkRusVer3 = 'Infantry AK ver3',
  InfantryM1Garand = 'soldier_wwii_us',
  InantryM249 = 'Soldier M249',
  InfantryM4 = 'Soldier M4',
  InfantryM4Georgia = 'Soldier M4 GRG',
  InfantryMauser98 = 'soldier_mauser98',
  InfantryRpg = 'Soldier RPG',
  InfantrySMLENo4Mk1 = 'soldier_wwii_br_01',
  InsurgentAk = 'Infantry AK Ins',
  JTAC = 'JTAC',
  ParatrooperAKS = 'Paratrooper AKS-74',
  ParatroperRPG = 'Paratrooper RPG-16',
}